// src/features/order/types/order.ts
export enum DeliveryType {
  PICKUP = "PICKUP", // El cliente retira en el local
  DELIVERY = "DELIVERY",
  IN_HOUSE_DELIVERY = "IN_HOUSE_DELIVERY", // El negocio entrega por su cuenta
  EXTERNAL_DELIVERY = "EXTERNAL_DELIVERY", // Se usa una cadetería externa
}

export enum PaymentMethodType {
  TRANSFER = "TRANSFER",
  CASH = "CASH",
  DELIVERY = "DELIVERY",
}

/// Estados de pago de una orden
export enum PaymentStatus {
  PENDING = "PENDING", /// Pago pendiente (cliente aún no inició)
  IN_PROGRESS = "IN_PROGRESS", /// Pago en curso (por ejemplo, transferencia en proceso)
  CONFIRMED = "CONFIRMED", /// Pago confirmado (negocio recibió el dinero)
  REJECTED = "REJECTED", /// Pago rechazado o fallido
}
// ...
export enum EOrderStatusBusiness {
  COMPLETED = "COMPLETED",
  WAITING_FOR_PAYMENT = "WAITING_FOR_PAYMENT", // Pedido esperando pago
  PENDING_CONFIRMATION = "PENDING_CONFIRMATION", // Esperando que el negocio acepte
  CONFIRMED = "CONFIRMED", // Negocio aceptó el pedido
  REJECTED_BY_BUSINESS = "REJECTED_BY_BUSINESS", // Negocio rechazó antes de preparar
  PREPARING = "PREPARING", // Pedido en preparación

  READY_FOR_CUSTOMER_PICKUP = "READY_FOR_CUSTOMER_PICKUP", // Pedido listo para retiro por el cliente
  READY_FOR_DELIVERY_PICKUP = "READY_FOR_DELIVERY_PICKUP", // Pedido listo y negocio llamó al delivery

  DELIVERY_PENDING = "DELIVERY_PENDING", // Buscando delivery para asignar
  DELIVERY_ASSIGNED = "DELIVERY_ASSIGNED", // Delivery asignado, esperando aceptación
  DELIVERY_REASSIGNING = "DELIVERY_REASSIGNING", // Buscando otro delivery tras rechazo

  CANCELLED_BY_BUSINESS = "CANCELLED_BY_BUSINESS", // Cancelado por el negocio
}